+++ pysui_gadgets/package/cmds/structs.py
@@ -34,22 +34,14 @@ def _struct_parm_types(holder: list, sig: str) -> str:
     return sig
 
 
-def _struct_parm_abilities(holder: list[SuiMoveStructTypeParameter], sig: str) -> str:
+def _struct_parm_abilities(holder: list[str], sig: str) -> str:
     """."""
     accum = []
     tindex = 0
     found = False
-    for struct_parm in holder:
-        if struct_parm.constraints.abilities:
-            found = True
-            accum.append(
-                f"T{tindex}: {' + '.join(struct_parm.constraints.abilities)}")
-            tindex += 1
-        else:
-            accum.append(f"T{tindex}")
-            tindex += 1
-    if found:
-        sig = f"<{', '.join(accum)}> "
+    accum = [x for x in holder]
+    if accum:
+        sig = f"<{','.join(accum)}> "
     return sig
 
 
@@ -78,24 +70,23 @@ def _field_signature(field: SuiMoveField, field_str: str) -> str:
 def print_module_structs(modules: dict[str, SuiMoveModule], args: Namespace) -> None:
     """."""
     if args.includes:
-        modules = filters.filter_include_modules(
-            mods=modules, includes=args.includes)
+        modules = filters.filter_include_modules(mods=modules, includes=args.includes)
     elif args.excludes:
-        modules = filters.filter_modules_excluding(
-            mods=modules, excludes=args.excludes)
+        modules = filters.filter_modules_excluding(mods=modules, excludes=args.excludes)
 
-    for mod_name, mod_def in modules.items():
-        print(f"\nModule {mod_name}")
-        for struct_name, struct_def in mod_def.structs.items():
-            sig = _struct_parm_abilities(struct_def.type_parameters, "")
-            sig += _struct_parm_types(struct_def.abilities.abilities, "")
-            print(f"\nStruct: {struct_name} {sig} {{")
-            max_field_name = max([len(x.name) for x in struct_def.fields])
+    for mod_def in modules:
+        print(f"\nModule {mod_def.module_name}")
+        for struct_def in mod_def.module_structures.structures:
+            sig = _struct_parm_abilities(struct_def.abilities, "")
+            # sig += _struct_parm_types(struct_def.abilities.abilities, "")
+            print(f"\nStruct: {struct_def.struct_name} {sig} {{")
+            max_field_name = max([len(x["field_name"]) for x in struct_def.fields])
             if not args.short_form:
                 for field in struct_def.fields:
-                    print(field.to_json(indent=2))
+                    print(field)
             else:
                 for field in struct_def.fields:
                     print(
-                        f"    {field.name:{max_field_name}s}: {_field_signature(field,'')}")
+                        f"    {field.name:{max_field_name}s}: {_field_signature(field,'')}"
+                    )
             print("}")