+++ pysui_gadgets/utils/filters.py
@@ -18,6 +18,7 @@ Provides low level filtering options of raw return of modules, structs and funct
 
 import functools
 from typing import Iterator
+import pysui.sui.sui_pgql.pgql_types as pgql_type
 from pysui.sui.sui_txresults.package_meta import (
     SuiMoveFunction,
     SuiMoveModule,
@@ -28,15 +29,16 @@ _MODULE_TUPLE_NAME_POS: int = 0
 _MODULE_TUPLE_FUNC_POS: int = 1
 _MODULE_TUPLE_MODULE_POS: int = 1
 
+
 # Predicates
-def module_name_is(include_list: list[str], mods: tuple[str, SuiMoveModule]) -> bool:
+def module_name_is(include_list: list[str], mods: pgql_type.MoveModuleGQL) -> bool:
     """."""
-    return mods[_MODULE_TUPLE_NAME_POS] in include_list
+    return mods.module_name in include_list
 
 
-def module_name_not(exclude_list: list[str], mods: tuple[str, SuiMoveModule]) -> bool:
+def module_name_not(exclude_list: list[str], mods: pgql_type.MoveModuleGQL) -> bool:
     """."""
-    return mods[_MODULE_TUPLE_NAME_POS] not in exclude_list
+    return mods.module_name not in exclude_list
 
 
 def all_functions(funcs: tuple[str, SuiMoveFunction]) -> bool:
@@ -51,10 +53,19 @@ def entry_functions_only(funcs: tuple[str, SuiMoveFunction]) -> bool:
 
 def mod_with_entry_points(mods: tuple[str, SuiMoveModule]) -> bool:
     """."""
-    return list((filter(entry_functions_only, mods[_MODULE_TUPLE_MODULE_POS].exposed_functions.items())))
+    return list(
+        (
+            filter(
+                entry_functions_only,
+                mods[_MODULE_TUPLE_MODULE_POS].exposed_functions.items(),
+            )
+        )
+    )
 
 
-def struct_abilities_with(abilities: set[str], structs: tuple[str, SuiMoveStruct]) -> bool:
+def struct_abilities_with(
+    abilities: set[str], structs: tuple[str, SuiMoveStruct]
+) -> bool:
     """."""
     return abilities.intersection(structs[1].abilities.abilities)
 
@@ -70,13 +81,13 @@ def filter_entry_points(module: SuiMoveModule) -> Iterator:
 
 
 def filter_modules_with_entry_points(
-    *, mods: dict[str, SuiMoveModule], nonentries: bool = False
+    *, mods: dict[str, pgql_type.MoveModuleGQL], nonentries: bool = False
 ) -> dict[str, list[dict[str, SuiMoveFunction]]]:
     """filter_entry_points Produces a list of functions found, and grouped, for one or more modules.
 
     :param mods: Dictionary of move modules
-    :type mods: dict[str, SuiMoveModule]
-    :param nonentries: Include both non-entry point and entry point functions for the module, defaults to False
+    :type mods: dict[str, pgql_type.MoveModuleGQL]
+    :param nonentries: Include both non-entry/non-public entry point functions for the module, defaults to False
     :type nonentries: bool, optional
     :return: A module key dictionary containing a function name key and function value dictionary
     :rtype: dict[str, list[dict[str, SuiMoveFunction]]]
@@ -85,29 +96,34 @@ def filter_modules_with_entry_points(
     result_dict = {}
     for mod_name, mod_def in mods.items():
         func_list = []
-        for func_name, func_def in mod_def.exposed_functions.items():
-            if func_def.is_entry or nonentries:
-                func_list.append({func_name: func_def})
-        if func_list:
-            result_dict[mod_name] = func_list
+        if mod_def.module_functions:
+            for func in mod_def.module_functions.functions:
+                if (func.is_entry or func.visibility == "PUBLIC") or nonentries:
+                    func_list.append({func.function_name: func})
+            if func_list:
+                result_dict[mod_name] = func_list
     return result_dict
 
 
-def filter_include_modules(*, mods: dict[str, SuiMoveModule], includes: list[str]) -> dict[str, SuiMoveModule]:
+def filter_include_modules(
+    *, mods: list[pgql_type.MoveModuleGQL], includes: list[str]
+) -> list[pgql_type.MoveModuleGQL]:
     """filter_include_modules Returns modules whose names are in the includes list.
 
     :param mods: SuiMovePackage modules dictionary
     :type mods: dict[str, SuiMoveModule]
     :param includes: List of module names to use in filter
     :type includes: list[str]
-    :return: Dictionary of module_name:SuiMoveModule that passed the includes critieria
-    :rtype: dict[str, SuiMoveModule]
+    :return: list of module_name:MoveModuleGQL that passed the includes critieria
+    :rtype: list[pgql_type.MoveModuleGQL]
     """
     # filter any includes only
-    return dict(filter(functools.partial(module_name_is, includes), mods.items()))
+    return list(filter(functools.partial(module_name_is, includes), mods))
 
 
-def filter_exclude_modules(*, mods: dict[str, SuiMoveModule], excludes: list[str]) -> dict[str, SuiMoveModule]:
+def filter_exclude_modules(
+    *, mods: dict[str, SuiMoveModule], excludes: list[str]
+) -> dict[str, SuiMoveModule]:
     """filter_exclude_modules Returns modules whose names are not in the excludes list.
 
     :param mods: SuiMovePackage modules dictionary
@@ -122,7 +138,10 @@ def filter_exclude_modules(*, mods: dict[str, SuiMoveModule], excludes: list[str
 
 
 def filter_modules_with_including(
-    *, mods: dict[str, SuiMoveModule], includes: list[str] = None, nonentries: bool = False
+    *,
+    mods: dict[str, pgql_type.MoveModuleGQL],
+    includes: list[str] = None,
+    nonentries: bool = False,
 ) -> dict[str, list[dict[str, SuiMoveFunction]]]:
     """filter_modules_with_including Produces a list of functions found, and grouped, for one or more modules.
 
@@ -144,7 +163,10 @@ def filter_modules_with_including(
 
 
 def filter_modules_excluding(
-    *, mods: dict[str, SuiMoveModule], excludes: list[str] = None, nonentries: bool = False
+    *,
+    mods: dict[str, SuiMoveModule],
+    excludes: list[str] = None,
+    nonentries: bool = False,
 ) -> dict[str, list[dict[str, SuiMoveFunction]]]:
     """filter_modules_excluding Produces a list of functions found, and grouped, for one or more modules.