+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - Unpublished
+
+**BREAKING** All gadgets have been updated to use Sui GraphQL in addition to command line changes.
+
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+- `dsl` and `module` commands
+
 ## [0.4.9] - 2024-05-03
 
 ### Added
@@ -14,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bumped requirements.txt for pysui to >= 0.49.0
+
+### Removed
 
 ## [0.4.3] - 2023-11-06
 
@@ -25,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bumped requirements.txt for pysui to >= 0.41.0
 
+### Removed
 
 ## [0.4.2] - 2023-11-06